move prior generation to its own file where we can include additional priors in the future
